@@ -5,11 +5,13 @@
 
 	<properties>
 		<java.version>17</java.version>
-		<spring-boot.version>3.0.5</spring-boot.version>
+		<spring-boot.version>3.1.0</spring-boot.version>
 		<json-web-token.version>0.11.5</json-web-token.version>
 		<lombok.version>1.18.22</lombok.version>
-		<liquibase.version>4.17.1</liquibase.version>
-		<auth0-jwt.version>4.2.1</auth0-jwt.version>
+		<liquibase.version>4.22.0</liquibase.version>
+		<auth0-jwt.version>4.3.0</auth0-jwt.version>
+		<mapstruct.version>1.5.5.Final</mapstruct.version>
+		<hibernate.version>6.2.3.Final</hibernate.version>
 	</properties>
 
 	<parent>
@@ -58,7 +60,7 @@
 		<dependency>
 			<groupId>com.google.code.gson</groupId>
 			<artifactId>gson</artifactId>
-			<version>2.9.0</version>
+			<version>2.10.1</version>
 		</dependency>
 
 		<!-- Auth0 -->
@@ -147,6 +149,18 @@
 			<version>3.0.1</version>
 		</dependency>
 
+		<dependency>
+			<groupId>org.hibernate.orm</groupId>
+			<artifactId>hibernate-core</artifactId>
+			<version>${hibernate.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>org.mapstruct</groupId>
+			<artifactId>mapstruct</artifactId>
+			<version>${mapstruct.version}</version>
+		</dependency>
+
 	</dependencies>
 
 	<build>
@@ -171,7 +185,7 @@
 						<path>
 							<groupId>org.mapstruct</groupId>
 							<artifactId>mapstruct-processor</artifactId>
-							<version>1.5.2.Final</version>
+							<version>${mapstruct.version}</version>
 						</path>
 						<path>
 							<groupId>org.projectlombok</groupId>
