@@ -1,17 +1,28 @@
 package reforged.marcin.krysiak.basketstats.config;
 
 import net.kaczmarzyk.spring.data.jpa.web.SpecificationArgumentResolver;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import reforged.marcin.krysiak.basketstats.interceptors.HeaderInterceptor;
 
 import java.util.List;
 
 @Configuration
-public class SpecificationConfiguration implements WebMvcConfigurer {
+public class WebApplicationConfig implements WebMvcConfigurer {
+
+    @Autowired
+    HeaderInterceptor headerInterceptor;
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
         argumentResolvers.add(new SpecificationArgumentResolver());
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(headerInterceptor);
     }
 }
