@@ -2,30 +2,30 @@ package reforged.marcin.krysiak.basketstats.controllers;
 
 
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import reforged.marcin.krysiak.basketstats.security.Message;
 
-
+/*
+ * Auth testing controller
+ */
 @RestController
 @RequestMapping("/api/messages")
 public class AuthController {
 
     @GetMapping(value = "/public")
-    public Message publicEndpoint() {
-        return new Message("All good. You DO NOT need to be authenticated to call /api/public.");
+    public String publicEndpoint() {
+        return "All good. You DO NOT need to be authenticated to call /api/public.";
     }
 
     @GetMapping(value = "/protected")
-    public Message privateEndpoint() {
-        return new Message("All good. You can see this because you are Authenticated.");
+    public String privateEndpoint() {
+        return "All good. You can see this because you are Authenticated.";
     }
 
     @GetMapping(value = "/admin")
     @PreAuthorize("hasAuthority('read:admin-messages')")
-    public Message privateScopedEndpoint() {
-        return new Message("All good. You can see this because you are Authenticated with a Token granted the 'read:admin-messages' scope");
+    public String privateScopedEndpoint() {
+        return "All good. You can see this because you are Authenticated with a Token granted the 'read:admin-messages' scope";
     }
 }
