@@ -20,7 +20,7 @@ public class Match {
 
     @Id
     @SequenceGenerator(allocationSize = 1, name = "matches_id_seq", sequenceName = "matches_id_seq")
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "matches_id_seq")
     @Column(name = "id")
     private Long id;
 
