@@ -1,7 +1,9 @@
 package reforged.marcin.krysiak.basketstats.models;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.util.Arrays;
@@ -10,58 +12,45 @@ import java.util.Arrays;
 @Setter
 @Entity
 @Table(name = "teams")
+@NoArgsConstructor
+@AllArgsConstructor
 public class Team {
 
     @Id
     @SequenceGenerator(allocationSize = 1, name = "teams_id_seq", sequenceName = "teams_id_seq")
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "teams_id_seq")
     private Long id;
 
     @Column(length = 50, nullable = false)
     private String name;
 
-    @Column(nullable = true)
-    private String logo;
+    @Column(name = "logo")
+    private String imageName;
 
-    @Column(nullable = true)
     @Basic(fetch = FetchType.LAZY)
-    private byte[] data;
+    @Column(name = "data")
+    private byte[] imageFile;
 
-    @Column(length = 100, nullable = true)
+    @Column(length = 100)
     private String type;
 
-    @ManyToOne()
-    @JoinColumn(name = "league_id", referencedColumnName = "id", nullable = false)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "league_id", nullable = false)
     private League league;
 
-    public Team() {
-
-    }
-
-    public Team(Long id, String name, String logo, League league) {
+    public Team(Long id, String name, String imageName, League league) {
         this.id = id;
         this.name = name;
-        this.logo = logo;
+        this.imageName = imageName;
         this.league = league;
     }
-
-    public Team(Long id, String name, String logo, byte[] data, String type, League league) {
-        this.id = id;
-        this.name = name;
-        this.logo = logo;
-        this.data = data;
-        this.type = type;
-        this.league = league;
-    }
-
     @Override
     public String toString() {
         return "Team{" +
                 "id=" + id +
                 ", name='" + name + '\'' +
-                ", logo='" + logo + '\'' +
-                ", data=" + Arrays.toString(data) +
+                ", logo='" + imageName + '\'' +
+                ", data=" + Arrays.toString(imageFile) +
                 ", type='" + type + '\'' +
                 ", league=" + league +
                 '}';
