@@ -1,13 +1,14 @@
 package reforged.marcin.krysiak.basketstats.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 import reforged.marcin.krysiak.basketstats.models.Team;
 
 import java.util.List;
 
 @Repository
-public interface TeamRepository extends JpaRepository<Team, Long> {
+public interface TeamRepository extends JpaRepository<Team, Long>, JpaSpecificationExecutor<Team> {
     List<Team> findAllByLeagueId(Long leagueId);
     Team findByName(String name);
 }
