@@ -21,6 +21,7 @@ import { MatTooltipModule } from "@angular/material/tooltip";
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AuthModule, AuthService } from "@auth0/auth0-angular";
+import { EffectsModule } from '@ngrx/effects';
 import { StoreRouterConnectingModule } from '@ngrx/router-store';
 import { StoreModule } from '@ngrx/store';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
@@ -38,6 +39,7 @@ import { NavHeaderComponent } from "./components/header/nav-header/nav-header.co
 import { SidebarHeaderComponent } from './components/header/sidebar-header/sidebar-header.component';
 import { HomepageComponent } from './components/homepage/homepage.component';
 import { LeagueComponent } from './components/league/league.component';
+import { LeagueEffects } from "./components/league/state/league.effects";
 import { MatchComponent } from './components/match/match.component';
 import { MatchModule } from "./components/match/match.module";
 import { MatchInfoRowComponent } from './components/schedule/match-info-row/match-info-row.component';
@@ -49,6 +51,7 @@ import { QuarterStatsForPlayersByTeamComponent } from './components/schedule/mat
 import { QuarterStatsForPlayersComponent } from './components/schedule/match-stats/match-stats-for-players/quarter-stats-for-players/quarter-stats-for-players.component';
 import { MatchStatsComponent } from './components/schedule/match-stats/match-stats.component';
 import { ScheduleComponent } from './components/schedule/schedule.component';
+import { TeamEffects } from "./components/team/state/team.effects";
 import { TeamPlayersAvgStatsListComponent } from './components/team/team-players-avg-stats-list/team-players-avg-stats-list.component';
 import { TeamPlayersListComponent } from './components/team/team-players-list/team-players-list.component';
 import { TeamPlayersComponent } from './components/team/team-players/team-players.component';
@@ -58,10 +61,9 @@ import { UserModule } from "./components/user/user.module";
 import { HttpAuthInterceptor } from "./helpers/interceptors/http-auth.interceptor";
 import { HttpErrorInterceptor } from "./helpers/interceptors/http-error.interceptor";
 import { SharedModule } from "./shared/shared.module";
+import { LoaderEffects } from './shared/state/effect/loader.effects';
 import { metaReducers, reducers } from './store';
 import { getPolishPaginatorIntl } from "./translations/polish-paginator-intl";
-import { EffectsModule } from '@ngrx/effects';
-import { LoaderEffects } from './shared/state/effect/loader.effects';
 
 registerLocaleData(localePl);
 
@@ -144,7 +146,11 @@ registerLocaleData(localePl);
     ),
     !environment.production ? StoreDevtoolsModule.instrument() : [],
     StoreRouterConnectingModule.forRoot(),
-    EffectsModule.forRoot([LoaderEffects])
+    EffectsModule.forRoot([
+      LoaderEffects,
+      LeagueEffects,
+      TeamEffects
+    ])
   ],
   providers: [
     AuthGuard,
