@@ -1,9 +1,8 @@
+import { ImageFile } from "../shared/model/image-file";
 import { League } from "./league.model";
 
-export class Team {
+export class Team extends ImageFile {
   id: number;
   name: string;
-  logo: string;
   league: League;
-  logoFile: Blob;
 }
