@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
 import { FormGroup, UntypedFormControl, Validators } from '@angular/forms';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { FieldType } from '../../enum/field-type';
@@ -37,7 +37,7 @@ export class BaseAddEditFormComponent<T extends BaseCrudResource> implements OnI
     this.formConfig.formFields.forEach(formFieldDef => {
       group[formFieldDef.name] = new UntypedFormControl(
         (this.dataObject as any)?.[formFieldDef.name],
-        [...formFieldDef.validators!]
+        !!formFieldDef.validators ? [...formFieldDef.validators!] : []
       );
     });
     return new FormGroup(group);
@@ -54,6 +54,10 @@ export class BaseAddEditFormComponent<T extends BaseCrudResource> implements OnI
         ...this.formGroup.getRawValue()
       }
     );
+  }
+
+  public compareWithId(object1: BaseCrudResource, object2: BaseCrudResource): boolean {
+    return !!object1?.id && !!object2?.id && object1.id === object2.id;
   }
 
   public get headerText(): string | undefined {
