@@ -3,7 +3,7 @@ import { TableOperation } from "../enum/table-operation";
 
 export interface TableConfig {
     tableColumns: TableColumn[];
-    headerText: string;
+    headerText?: string;
     searchPlaceholder: string;
     addButtonTooltip: string;
     pageSizeOptions?: number[];
@@ -14,4 +14,16 @@ export interface TableColumn {
     name: string;
     type: ColumnType;
     translation?: string;
+}
+
+export function getCopyWithColumnIfNotExists(tableColumns: TableColumn[], column: TableColumn, first?: boolean) {
+    tableColumns = !!tableColumns ? tableColumns?.filter(tableColumn => tableColumn.name !== column.name) : [];
+
+    if (first) {
+        tableColumns.unshift(column);
+    } else {
+        tableColumns.push(column);
+    }
+
+    return tableColumns;
 }
