@@ -3,14 +3,14 @@ import {
   ActionReducerMap, MetaReducer
 } from '@ngrx/store';
 import { environment } from 'src/environments/environment';
-// import * as fromLeague from './league/league.reducer';
 import * as fromLeague from '../components/league/state/league.reducer';
+import * as fromTeam from '../components/team/state/team.reducer';
 
 
 export interface BasketStatsAppState {
 
-  // [fromLeague.leagueFeatureKey]: fromLeague.LeagueState;
   [fromLeague.leaguesFeatureKey]: fromLeague.LeagueState;
+  [fromTeam.teamsFeatureKey]: fromTeam.TeamState;
 
   router: fromRouter.RouterReducerState;
 
@@ -18,8 +18,8 @@ export interface BasketStatsAppState {
 
 export const reducers: ActionReducerMap<BasketStatsAppState> = {
 
-  // [fromLeague.leagueFeatureKey]: fromLeague.reducer,
   [fromLeague.leaguesFeatureKey]: fromLeague.reducer,
+  [fromTeam.teamsFeatureKey]: fromTeam.reducer,
 
   router: fromRouter.routerReducer
 
